# el-pasas/wwwroot/swagger-original.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "",
    "version" : "0.1",
    "title" : "eDocument API",
    "termsOfService" : ""
  },
  "host" : "el-pasas.azurewebsites.net",
  "basePath" : "/api",
...
  "security" : [ {
    "Token" : [ ]
  } ],
  "paths" : {
...
    "/passport/{userId}" : {
      "get" : {
        "tags" : [ "Passport" ],
        "summary" : "Get Passport for the User",
        "operationId" : "getPassport",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully retrieved Passport model",
            "schema" : {
              "$ref" : "#/definitions/Passport"
            }
          },
          "404" : {
            "description" : "Passport for this {userId} not found"
          }
        }
      }
    },
    "/passport/{userId}/image/small" : {
      "get" : {
        "tags" : [ "Passport" ],
        "summary" : "Get small profile image for the Passport which belongs to the User",
        "operationId" : "getPassportImageSmall",
        "produces" : [ "image/jpeg" ],
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully retrieved Passport small profile image"
          },
          "404" : {
            "description" : "Passport small image for this {userId} not found"
          }
        }
      }
    },
    "/passport/{userId}/signature" : {
      "get" : {
        "tags" : [ "Passport" ],
        "summary" : "Get signature image for the Passport which belongs to the User",
        "operationId" : "getPassportSignature",
        "produces" : [ "image/jpeg" ],
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully retrieved Passport signature image"
          },
          "404" : {
            "description" : "Passport signature image for this {userId} not found"
          }
        }
      }
    },
    "/passport/{userId}/qr" : {
      "get" : {
        "tags" : [ "Passport" ],
        "summary" : "Get QR code for the Passport which belongs to the User",
        "operationId" : "getPassportQr",
        "produces" : [ "text/plain" ],
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully retrieved Passport QR code"
          },
          "404" : {
            "description" : "Passport QR code for this {userId} not found"
          }
        }
      }
    },
    "/passport/{userId}/nfc" : {
      "get" : {
        "tags" : [ "Passport" ],
        "summary" : "Get NFC code for the Passport which belongs to the User",
        "operationId" : "getPassportNfc",
        "produces" : [ "text/plain" ],
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully retrieved Passport NFC code"
          },
          "404" : {
            "description" : "Passport NFC code for this {userId} not found"
          }
        }
      }
    },
    "/passport/{userId}/image" : {
      "get" : {
        "tags" : [ "Passport" ],
        "summary" : "Get profile image for the Passport which belongs to the User",
        "operationId" : "getPassportImage",
        "produces" : [ "image/jpeg" ],
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully retrieved Passport profile image"
          },
          "404" : {
            "description" : "Passport image for this {userId} not found"
          }
        }
      },
      "put" : {
        "tags" : [ "Passport" ],
        "summary" : "Update profile image for the Passport which belongs to the User",
        "operationId" : "putPassportImage",
        "consumes" : [ "multipart/form-data" ],
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "image",
          "in" : "formData",
          "description" : "Image to upload",
          "required" : true,
          "type" : "file"
        } ],
        "responses" : {
          "204" : {
            "description" : "Successfully uploaded Passport profile image"
          },
          "400" : {
            "description" : "Provided image is invalid"
          }
        }
      }
    },
    "/passport/{userId}/validate" : {
      "put" : {
        "tags" : [ "Passport" ],
        "summary" : "Request to revalidate and extend expiration for Passport",
        "operationId" : "putPassportValidation",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "204" : {
            "description" : "Successfully requested Passport revalidation"
          },
          "404" : {
            "description" : "Passport for this {userId} not found"
          }
        }
      }
    },
...
  },
  "securityDefinitions" : {
    "Token" : {
      "type" : "basic"
    }
  },
  "definitions" : {
    "Passport" : {
      "type" : "object",
      "properties" : {
        "passportNo" : {
          "type" : "integer",
          "example" : 14951794
        },
        "surname" : {
          "type" : "string",
          "example" : "Basanavičienė"
        },
        "givenName" : {
          "type" : "string",
          "example" : "Birutė"
        },
        "nationality" : {
          "type" : "string",
          "example" : "Lietuvos Respublikos"
        },
        "dateOfBirth" : {
          "type" : "string",
          "format" : "dd mm yyyy",
          "example" : "11 03 1990"
        },
        "sex" : {
          "type" : "string",
          "example" : "MOT/F"
        },
        "dateOfIssue" : {
          "type" : "string",
          "format" : "dd mm yyyy",
          "example" : "02 01 2019"
        },
        "dateOfExpiry" : {
          "type" : "string",
          "format" : "dd mm yyyy",
          "example" : "02 01 2029"
        },
        "personalNo" : {
          "type" : "integer",
          "example" : 49003111045
        },
        "placeOfBirth" : {
          "type" : "string",
          "example" : "Lietuva"
        },
        "authority" : {
          "type" : "string",
          "example" : "Vilnius (19)"
        },
        "footer" : {
          "type" : "string",
          "example" : "P<LTUBASANAVICIENE<<BIRUTE<<<<<<<<<<<<<<<<<<14951794<0LTU90031187290102049003111045<<<86",
          "minLength" : 88,
          "maxLength" : 88
        }
      },
      "example" : {
        "passportNo" : 14951794,
        "dateOfExpiry" : "02 01 2029",
        "placeOfBirth" : "Lietuva",
        "nationality" : "Lietuvos Respublikos",
        "footer" : "P<LTUBASANAVICIENE<<BIRUTE<<<<<<<<<<<<<<<<<<14951794<0LTU90031187290102049003111045<<<86",
        "surname" : "Basanavičienė",
        "givenName" : "Birutė",
        "sex" : "MOT/F",
        "authority" : "Vilnius (19)",
        "personalNo" : 49003111045,
        "dateOfBirth" : "11 03 1990",
        "dateOfIssue" : "02 01 2019"
      }
    },
...
  }
}